Mounted Labyrinth in Home.jsx.
